pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "lightrag-mcp"
version = "0.1.0"
description = "MCP Server для LightRAG"
requires-python = ">=3.10"
readme = "README.md"
authors = [
    {name = "Author", email = "author@example.com"},
]
dependencies = [
    "mcp>=0.2.0",
    "httpx>=0.25.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "mypy>=1.5.0",
    "black>=23.7.0",
    "isort>=5.12.0",
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.1",
]

[project.scripts]
lightrag-mcp = "lightrag_mcp.main:main"

[tool.hatch.build.targets.wheel]
packages = ["src/lightrag_mcp"]
